cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lich-framework"
version = "1.0.0"
description = "AI-Ready Full-Stack Project Generator"
license = {text = "MIT"}
authors = [
    { name = "Babak", email = "babak@example.com" }
]
keywords = ["cookiecutter", "project-generator", "fastapi", "nextjs", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "cookiecutter>=2.5.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
lich = "lich.cli:app"

[project.urls]
Homepage = "https://github.com/DoTech/lich"

[tool.setuptools.packages.find]
where = ["src"]
